gn: Cleanup chromium scripts
Move gn.py into scripts/gn

Minor script cleanup from Vulkan-Loader
